requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
openai>=1.12.0
pydantic>=2.6.0
typing-extensions>=4.9.0
python-multipart>=0.0.9
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
numpy>=1.26.0
python-firebase>=1.2.1
python-jose>=3.3.0
huggingface-hub>=0.20.0
transformers>=4.37.0
faiss-cpu>=1.7.4
tiktoken>=0.6.0 
plotly>=5.0.0